NXP-28818: Lower CPU consumption in RuntimeFeature initialization
By lowering the number of threads used for the recursive classpath scan to a fixed number, 2, instead of computing a ratio of the available processors.

This is required when running the unit tests in parallel in the same container, otherwise the number of threads can grow very fast and become unreasonable regarding the available CPU resources.

For instance, let's say we are running in a container with 8 CPUs.
The number of threads computed by ClassGraph#DEFAULT_NUM_WORKER_THREADS will be 4 + 10 = 14.
- If we're running 1 Maven process, we'll have 14 threads.
- If we're running 3 Maven processes, we'll have 42 threads! This leads to BLOCKED threads, thus probably causes some latency.